+++ InvestmentManager/InvestMgr.h
@@ -18,7 +18,7 @@ public:
 	virtual bool GetData(std::wstring agentId, std::wstring dataId,
 		int year, int month, int day, CDataItem& data);
 	virtual bool GetData(std::wstring agentId, std::wstring dataId,
-		int fromYear, int fromMonth, int fromDay, int dayCount, std::vector<CDataItem>& data);
+		int fromYear, int fromMonth, int fromDay, size_t dayCount, std::vector<CDataItem>& data);
 	virtual bool GetSimpleData(std::wstring agentId, std::vector<std::wstring> dataId,
 		int year, int month, int day, std::vector<std::wstring>& data);
 	virtual bool GetSimpleData(std::wstring agentId, std::wstring dataId,
